Change Nip19 from class to object
