+++ app/src/main/java/com/vitorpamplona/amethyst/ui/components/ClickableRoute.kt
@@ -10,8 +10,8 @@ import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.text.AnnotatedString
 import androidx.navigation.NavController
 import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.service.Nip19
 import com.vitorpamplona.amethyst.service.model.ChannelCreateEvent
+import com.vitorpamplona.amethyst.service.nip19.Nip19
 
 @Composable
 fun ClickableRoute(
