+++ app/src/test/java/com/vitorpamplona/amethyst/service/Nip19Test.kt
@@ -1,46 +1,21 @@
 package com.vitorpamplona.amethyst.service
 
+import com.vitorpamplona.amethyst.service.nip19.Nip19
 import org.junit.Assert
 import org.junit.Ignore
 import org.junit.Test
 
 class Nip19Test {
-
-    private val nip19 = Nip19()
-
-    @Test(expected = IllegalArgumentException::class)
-    fun to_int_32_length_smaller_than_4() {
-        toInt32(byteArrayOfInts(1, 2, 3))
-    }
-
-    @Test(expected = IllegalArgumentException::class)
-    fun to_int_32_length_bigger_than_4() {
-        toInt32(byteArrayOfInts(1, 2, 3, 4, 5))
-    }
-
-    @Test()
-    fun to_int_32_length_4() {
-        val actual = toInt32(byteArrayOfInts(1, 2, 3, 4))
-
-        Assert.assertEquals(16909060, actual)
-    }
-
-    @Ignore("Test not implemented yet")
-    @Test()
-    fun parse_TLV() {
-        // TODO: I don't know how to test this (?)
-    }
-
     @Test()
     fun uri_to_route_null() {
-        val actual = nip19.uriToRoute(null)
+        val actual = Nip19.uriToRoute(null)
 
         Assert.assertEquals(null, actual)
     }
 
     @Test()
     fun uri_to_route_unknown() {
-        val actual = nip19.uriToRoute("nostr:unknown")
+        val actual = Nip19.uriToRoute("nostr:unknown")
 
         Assert.assertEquals(null, actual)
     }
@@ -48,7 +23,7 @@ class Nip19Test {
     @Test()
     fun uri_to_route_npub() {
         val actual =
-            nip19.uriToRoute("nostr:npub1hv7k2s755n697sptva8vkh9jz40lzfzklnwj6ekewfmxp5crwdjs27007y")
+            Nip19.uriToRoute("nostr:npub1hv7k2s755n697sptva8vkh9jz40lzfzklnwj6ekewfmxp5crwdjs27007y")
 
         Assert.assertEquals(Nip19.Type.USER, actual?.type)
         Assert.assertEquals(
@@ -60,7 +35,7 @@ class Nip19Test {
     @Test()
     fun uri_to_route_note() {
         val actual =
-            nip19.uriToRoute("nostr:note1stqea6wmwezg9x6yyr6qkukw95ewtdukyaztycws65l8wppjmtpscawevv")
+            Nip19.uriToRoute("nostr:note1stqea6wmwezg9x6yyr6qkukw95ewtdukyaztycws65l8wppjmtpscawevv")
 
         Assert.assertEquals(Nip19.Type.NOTE, actual?.type)
         Assert.assertEquals(
@@ -72,7 +47,7 @@ class Nip19Test {
     @Ignore("Test not implemented yet")
     @Test()
     fun uri_to_route_nprofile() {
-        val actual = nip19.uriToRoute("nostr:nprofile")
+        val actual = Nip19.uriToRoute("nostr:nprofile")
 
         Assert.assertEquals(Nip19.Type.USER, actual?.type)
         Assert.assertEquals("*", actual?.hex)
@@ -81,7 +56,7 @@ class Nip19Test {
     @Ignore("Test not implemented yet")
     @Test()
     fun uri_to_route_nevent() {
-        val actual = nip19.uriToRoute("nostr:nevent")
+        val actual = Nip19.uriToRoute("nostr:nevent")
 
         Assert.assertEquals(Nip19.Type.USER, actual?.type)
         Assert.assertEquals("*", actual?.hex)
@@ -90,7 +65,7 @@ class Nip19Test {
     @Ignore("Test not implemented yet")
     @Test()
     fun uri_to_route_nrelay() {
-        val actual = nip19.uriToRoute("nostr:nrelay")
+        val actual = Nip19.uriToRoute("nostr:nrelay")
 
         Assert.assertEquals(Nip19.Type.RELAY, actual?.type)
         Assert.assertEquals("*", actual?.hex)
@@ -99,11 +74,9 @@ class Nip19Test {
     @Ignore("Test not implemented yet")
     @Test()
     fun uri_to_route_naddr() {
-        val actual = nip19.uriToRoute("nostr:naddr")
+        val actual = Nip19.uriToRoute("nostr:naddr")
 
         Assert.assertEquals(Nip19.Type.ADDRESS, actual?.type)
         Assert.assertEquals("*", actual?.hex)
     }
-
-    private fun byteArrayOfInts(vararg ints: Int) = ByteArray(ints.size) { pos -> ints[pos].toByte() }
 }
